Create apierrors package, improve handling of password and email validation and the append function
