+++ api/tokens.go
@@ -7,25 +7,21 @@ import (
 	"io/ioutil"
 	"net/http"
 
+	"github.com/ONSdigital/dp-identity-api/apierrors"
 	"github.com/ONSdigital/dp-identity-api/emailvalidation"
 
 	"github.com/ONSdigital/log.go/log"
 )
 
-type ErrorStructure struct {
-	Errors []IndividualError `json:"errors"`
+type AuthParams struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
 }
 
-type IndividualError struct {
-	SpecificError string `json:"error"`
-	Message       string `json:"message"`
-	Source        Source `json:"source"`
-}
-
-type Source struct {
-	Field string `json:"field"`
-	Param string `json:"param"`
-}
+var invalidPasswordError = errors.New("Invalid password")
+var invalidPasswordMessage = "Unable to validate the password in the request"
+var invalidEmailError = errors.New("Invalid email")
+var invalidErrorMessage = "Unable to validate the email in the request"
 
 func TokensHandler() http.HandlerFunc {
 
@@ -36,6 +32,8 @@ func TokensHandler() http.HandlerFunc {
 		field := ""
 		param := ""
 
+		defer req.Body.Close()
+
 		body, err := ioutil.ReadAll(req.Body)
 		if err != nil {
 			errorMessage := "api endpoint POST login returned an error reading the request body"
@@ -43,9 +41,7 @@ func TokensHandler() http.HandlerFunc {
 			return
 		}
 
-		defer req.Body.Close()
-
-		authParams := make(map[string]string)
+		var authParams AuthParams
 		err = json.Unmarshal(body, &authParams)
 		if err != nil {
 			errorMessage := "api endpoint POST login returned an error unmarshalling the body"
@@ -61,89 +57,36 @@ func TokensHandler() http.HandlerFunc {
 			return
 		}
 
-		invalidPasswordError := errors.New("Invalid password")
-		invalidPasswordMessage := "Unable to validate the password in the request"
-		invalidEmailError := errors.New("Invalid email")
-		invalidErrorMessage := "Unable to validate the email in the request"
+		invalidPasswordErrorBody := apierrors.IndividualErrorBuilder(invalidPasswordError, invalidPasswordMessage, field, param)
+		invalidEmailErrorBody := apierrors.IndividualErrorBuilder(invalidEmailError, invalidErrorMessage, field, param)
 
-		invalidPasswordErrorBody := individualErrorBuilder(invalidPasswordError, invalidPasswordMessage, field, param)
-		invalidEmailErrorBody := individualErrorBuilder(invalidEmailError, invalidErrorMessage, field, param)
-
-		var errorList []IndividualError
-
-		if !(validPasswordRequest) && !(validEmailRequest) {
-
-			errorList = nil
-			errorList = append(errorList, invalidPasswordErrorBody)
-			errorList = append(errorList, invalidEmailErrorBody)
-
-			errorResponseBody := errorResponseBodyBuilder(errorList)
-			writeErrorResponse(ctx, w, 400, errorResponseBody)
-			return
-		}
+		var errorList []apierrors.IndividualError
 
 		if !validPasswordRequest {
-
-			errorList = nil
 			errorList = append(errorList, invalidPasswordErrorBody)
-			errorResponseBody := errorResponseBodyBuilder(errorList)
-
-			writeErrorResponse(ctx, w, 400, errorResponseBody)
-			return
 		}
 
 		if !validEmailRequest {
-
-			errorList = nil
 			errorList = append(errorList, invalidEmailErrorBody)
-			errorResponseBody := errorResponseBodyBuilder(errorList)
-
-			writeErrorResponse(ctx, w, 400, errorResponseBody)
-			return
 		}
+
+		errorResponseBody := apierrors.ErrorResponseBodyBuilder(errorList)
+		writeErrorResponse(ctx, w, 400, errorResponseBody)
+		return
 	}
 }
 
-func passwordValidation(requestBody map[string]string) (isPasswordValid bool) {
+func passwordValidation(requestBody AuthParams) (isPasswordValid bool) {
 
-	isPasswordValid = false
-
-	if len(requestBody["password"]) != 0 {
-		isPasswordValid = true
-	}
-
-	return isPasswordValid
+	return len(requestBody.Password) > 0
 }
 
 //emailValidation checks for both a valid email address and an empty email address
-func emailValidation(requestBody map[string]string) (isEmailValid bool) {
+func emailValidation(requestBody AuthParams) (isEmailValid bool) {
 
-	isEmailValid = false
-	isEmailValid = emailvalidation.IsEmailValid(requestBody["email"])
+	isEmailValid = emailvalidation.IsEmailValid(requestBody.Email)
 
 	return isEmailValid
-}
-
-func individualErrorBuilder(err error, message, sourceField, sourceParam string) (individualError IndividualError) {
-
-	individualError = IndividualError{
-		SpecificError: error.Error(err),
-		Message:       message,
-		Source: Source{
-			Field: sourceField,
-			Param: sourceParam},
-	}
-
-	return individualError
-}
-
-func errorResponseBodyBuilder(listOfErrors []IndividualError) (errorResponseBody ErrorStructure) {
-
-	errorResponseBody = ErrorStructure{
-		Errors: listOfErrors,
-	}
-
-	return errorResponseBody
 }
 
 func writeErrorResponse(ctx context.Context, w http.ResponseWriter, status int, errorResponseBody interface{}) {
@@ -151,27 +94,29 @@ func writeErrorResponse(ctx context.Context, w http.ResponseWriter, status int, 
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(status)
 
-	jsonResponse, _ := json.Marshal(errorResponseBody)
-	_, err := w.Write(jsonResponse)
+	jsonResponse, err := json.Marshal(errorResponseBody)
 	if err != nil {
-
-		log.Event(ctx, "writing response failed", log.Error(err), log.ERROR)
-		http.Error(w, "Failed to write http response", http.StatusInternalServerError)
+		log.Event(ctx, "failed to marshal the error", log.Error(err), log.ERROR)
+		http.Error(w, "failed to marshal the error", http.StatusInternalServerError)
 		return
 	}
 
-	return
+	_, err = w.Write(jsonResponse)
+	if err != nil {
+		log.Event(ctx, "writing response failed", log.Error(err), log.ERROR)
+		http.Error(w, "failed to write http response", http.StatusInternalServerError)
+		return
+	}
 }
 
 func handleUnexpectedError(ctx context.Context, w http.ResponseWriter, err error, message, sourceField, sourceParam string) {
 
-	var errorList []IndividualError
+	var errorList []apierrors.IndividualError
 
-	errorList = nil
 	statusCode := 500
-	internalServerErrorBody := individualErrorBuilder(err, message, sourceField, sourceParam)
+	internalServerErrorBody := apierrors.IndividualErrorBuilder(err, message, sourceField, sourceParam)
 	errorList = append(errorList, internalServerErrorBody)
-	errorResponseBody := errorResponseBodyBuilder(errorList)
+	errorResponseBody := apierrors.ErrorResponseBodyBuilder(errorList)
 
 	log.Event(ctx, message, log.ERROR, log.Error(err))
 	writeErrorResponse(ctx, w, statusCode, errorResponseBody)
