+++ api/tokens_test.go
@@ -7,16 +7,18 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/ONSdigital/dp-identity-api/apierrors"
+
 	. "github.com/smartystreets/goconvey/convey"
 )
 
 func TestPasswordHasBeenProvided(t *testing.T) {
 
 	Convey("A password has been provided", t, func() {
-		password := "password"
 
-		body := make(map[string]string)
-		body["password"] = password
+		body := AuthParams{
+			Password: "password",
+		}
 
 		passwordResponse := passwordValidation(body)
 
@@ -25,7 +27,7 @@ func TestPasswordHasBeenProvided(t *testing.T) {
 
 	Convey("There isn't a password field in the body and the password isn't validated", t, func() {
 
-		body := make(map[string]string)
+		body := AuthParams{}
 
 		passwordResponse := passwordValidation(body)
 		So(passwordResponse, ShouldBeFalse)
@@ -33,9 +35,9 @@ func TestPasswordHasBeenProvided(t *testing.T) {
 
 	Convey("There isn't a password value in the body and the password isn't validated", t, func() {
 
-		password := ""
-		body := make(map[string]string)
-		body["password"] = password
+		body := AuthParams{
+			Password: "",
+		}
 
 		passwordResponse := passwordValidation(body)
 		So(passwordResponse, ShouldBeFalse)
@@ -46,10 +48,9 @@ func TestEmailConformsToExpectedFormat(t *testing.T) {
 
 	Convey("The email conforms to the expected format and is validated", t, func() {
 
-		email := "email.email@ons.gov.uk"
-
-		body := make(map[string]string)
-		body["email"] = email
+		body := AuthParams{
+			Email: "email.email@ons.gov.uk",
+		}
 
 		emailResponse := emailValidation(body)
 		So(emailResponse, ShouldBeTrue)
@@ -57,9 +58,9 @@ func TestEmailConformsToExpectedFormat(t *testing.T) {
 
 	Convey("There isn't an email field in the body and the email isn't validated", t, func() {
 
-		password := "password"
-		body := make(map[string]string)
-		body["password"] = password
+		body := AuthParams{
+			Password: "password",
+		}
 
 		emailResponse := emailValidation(body)
 		So(emailResponse, ShouldBeFalse)
@@ -67,9 +68,9 @@ func TestEmailConformsToExpectedFormat(t *testing.T) {
 
 	Convey("There isn't a email value in the body and the email isn't validated", t, func() {
 
-		email := ""
-		body := make(map[string]string)
-		body["email"] = email
+		body := AuthParams{
+			Email: "",
+		}
 
 		emailResponse := emailValidation(body)
 		So(emailResponse, ShouldBeFalse)
@@ -77,67 +78,12 @@ func TestEmailConformsToExpectedFormat(t *testing.T) {
 
 	Convey("The email doesn't conform to the expected format and it isn't validated", t, func() {
 
-		email := "email"
-
-		body := make(map[string]string)
-		body["email"] = email
+		body := AuthParams{
+			Email: "email",
+		}
 
 		emailResponse := emailValidation(body)
 		So(emailResponse, ShouldBeFalse)
-	})
-}
-
-func TestBuildingIndividualErrors(t *testing.T) {
-
-	Convey("The individual error conforms to the expected structure", t, func() {
-
-		err := errors.New("string, unchanging so devs can use this in code")
-		message := "detailed explanation of error"
-		sourceField := "reference to field like some.field or something"
-		sourceParam := "query param causing issue"
-
-		individualErrorExample := IndividualError{
-			SpecificError: "string, unchanging so devs can use this in code",
-			Message:       "detailed explanation of error",
-			Source: Source{
-				Field: "reference to field like some.field or something",
-				Param: "query param causing issue"},
-		}
-
-		individualError := individualErrorBuilder(err, message, sourceField, sourceParam)
-
-		So(individualError, ShouldResemble, individualErrorExample)
-
-	})
-}
-
-func TestBuildingErrorStructure(t *testing.T) {
-	Convey("An error structure is created from a list of errors", t, func() {
-
-		listOfErrors := []IndividualError{
-			{
-				SpecificError: "string, unchanging so devs can use this in code",
-				Message:       "detailed explanation of error",
-				Source: Source{
-					Field: "reference to field like some.field or something",
-					Param: "query param causing issue"},
-			},
-		}
-
-		errorResponseBodyExample := ErrorStructure{
-			Errors: []IndividualError{
-				{
-					SpecificError: "string, unchanging so devs can use this in code",
-					Message:       "detailed explanation of error",
-					Source: Source{
-						Field: "reference to field like some.field or something",
-						Param: "query param causing issue"}},
-			},
-		}
-
-		errorResponseBody := errorResponseBodyBuilder(listOfErrors)
-
-		So(errorResponseBody, ShouldResemble, errorResponseBodyExample)
 	})
 }
 
@@ -146,21 +92,21 @@ func TestWriteErrorResponse(t *testing.T) {
 
 		errorResponseBodyExample := `{"errors":[{"error":"Invalid email","message":"Unable to validate the email in the request","source":{"field":"","param":""}},{"error":"Invalid email","message":"Unable to validate the email in the request","source":{"field":"","param":""}}]}`
 
-		var errorList []IndividualError
+		var errorList []apierrors.IndividualError
 		errorList = nil
 
 		invalidEmailError := errors.New("Invalid email")
 		invalidErrorMessage := "Unable to validate the email in the request"
 		field := ""
 		param := ""
-		invalidEmailErrorBody := individualErrorBuilder(invalidEmailError, invalidErrorMessage, field, param)
+		invalidEmailErrorBody := apierrors.IndividualErrorBuilder(invalidEmailError, invalidErrorMessage, field, param)
 		errorList = append(errorList, invalidEmailErrorBody)
 		errorList = append(errorList, invalidEmailErrorBody)
 
 		ctx := context.Background()
 		resp := httptest.NewRecorder()
 		statusCode := 400
-		errorResponseBody := errorResponseBodyBuilder(errorList)
+		errorResponseBody := apierrors.ErrorResponseBodyBuilder(errorList)
 
 		writeErrorResponse(ctx, resp, statusCode, errorResponseBody)
