+++ apierrors/apierrors.go
@@ -1,0 +1,38 @@
+package apierrors
+
+type ErrorStructure struct {
+	Errors []IndividualError `json:"errors"`
+}
+
+type IndividualError struct {
+	SpecificError string `json:"error"`
+	Message       string `json:"message"`
+	Source        Source `json:"source"`
+}
+
+type Source struct {
+	Field string `json:"field"`
+	Param string `json:"param"`
+}
+
+func IndividualErrorBuilder(err error, message, sourceField, sourceParam string) (individualError IndividualError) {
+
+	individualError = IndividualError{
+		SpecificError: error.Error(err),
+		Message:       message,
+		Source: Source{
+			Field: sourceField,
+			Param: sourceParam},
+	}
+
+	return individualError
+}
+
+func ErrorResponseBodyBuilder(listOfErrors []IndividualError) (errorResponseBody ErrorStructure) {
+
+	errorResponseBody = ErrorStructure{
+		Errors: listOfErrors,
+	}
+
+	return errorResponseBody
+}
